Refactor Security and User Role Management
- Organized imports in private_ai_keys.py and users.py for better readability.
- Updated role validation logic to utilize UserRole.get_all_roles() for consistency across user role checks.
- Removed unused role definitions and hierarchy from security.py, streamlining the codebase.
- Enhanced role management by centralizing role-related functionality in the roles module.
