+++ app/api/private_ai_keys.py
@@ -13,9 +13,22 @@ from app.schemas.models import (
 from app.db.postgres import PostgresManager
 from app.db.models import DBPrivateAIKey, DBRegion, DBUser, DBTeam
 from app.services.litellm import LiteLLMService
-from app.core.security import get_current_user_from_auth, get_role_min_key_creator, get_role_min_team_admin, get_private_ai_access, UserRole, check_system_admin
+from app.core.security import (
+    get_current_user_from_auth,
+    get_role_min_team_admin,
+    get_private_ai_access,
+    check_system_admin
+)
+from app.core.roles import UserRole
 from app.core.config import settings
-from app.core.resource_limits import check_key_limits, check_vector_db_limits, get_token_restrictions, DEFAULT_KEY_DURATION, DEFAULT_MAX_SPEND, DEFAULT_RPM_PER_KEY
+from app.core.resource_limits import (
+    check_key_limits,
+    check_vector_db_limits,
+    get_token_restrictions,
+    DEFAULT_KEY_DURATION,
+    DEFAULT_MAX_SPEND,
+    DEFAULT_RPM_PER_KEY
+)
 
 router = APIRouter(
     tags=["private-ai-keys"]
